Implement responsive design for FavoritesGrid component
- Updated FavoritesGrid.jsx to dynamically adjust the number of Pokémon cards per row based on screen size.
  - Uses a combination of Flexbox and CSS Grid layout to ensure cards are displayed optimally on different devices.
- Added media queries in FavoritesGrid.css to handle various screen sizes:
  - Default to 3 cards per row on large screens.
  - Adjusts to 2 cards per row on medium screens.
  - Adjusts to 1 card per row on small screens.
- Integrated Pagination component for navigation through pages of favorite Pokémon.
- Improved overall user experience on smaller screens by setting card width to 90%.
