+++ src/App.jsx
@@ -1,7 +1,6 @@
 import "./App.css";
 import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
-import NavBar from "./components/NavBar";
 import MainContent from './components/MainContent';
 
 function App() {
