+++ src/assets/typeColors.js
@@ -1,0 +1,24 @@
+// src/assets/typeColors.js
+
+const typeColors = {
+    bug: "#26de81",
+    dragon: "#ffeaa7",
+    electric: "#fed330",
+    fairy: "#FF0069",
+    fighting: "#30336b",
+    fire: "#f0932b",
+    flying: "#81ecec",
+    grass: "#00b894",
+    ground: "#EFB549",
+    ghost: "#a55eea",
+    ice: "#98D8D8",
+    normal: "#95afc0",
+    poison: "#6c5ce7",
+    psychic: "#a29bfe",
+    rock: "#b8a038",
+    steel: "#5a8ea2", 
+    dark: "#705746", 
+    water: "#0190FF",
+  };
+
+  export default typeColors;
