+++ src/components/NavBar.jsx
@@ -1,5 +1,5 @@
 // src/components/Navbar.jsx
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
 import pokeball from '../assets/pokeball.svg';
