+++ src/components/Pagination.jsx
@@ -1,0 +1,33 @@
+// src/components/Pagination.jsx
+import React from 'react';
+
+function Pagination({ currentPage, totalPages, handlePageChange , main }) {
+  const pageNumbers = [];
+  const pagesToShow = 5;
+  const startPage = Math.max(1, currentPage - Math.floor(pagesToShow / 2));
+  const endPage = Math.min(totalPages, startPage + pagesToShow - 1);
+
+  for (let i = startPage; i <= endPage; i++) {
+    pageNumbers.push(i);
+  }
+
+  return (
+    <nav aria-label="Page navigation" className={` ${main ? 'mt-3 d-flex justify-content-center' : ''}`}>
+      <ul className="pagination">
+        <li className={`page-item ${currentPage === 1 ? 'disabled' : ''}`}>
+          <button className="page-link" onClick={() => handlePageChange(currentPage - 1)}>&lt;</button>
+        </li>
+        {pageNumbers.map(page => (
+          <li key={page} className={`page-item ${page === currentPage ? 'active' : ''}`}>
+            <button className="page-link" onClick={() => handlePageChange(page)}>{page}</button>
+          </li>
+        ))}
+        <li className={`page-item ${currentPage === totalPages ? 'disabled' : ''}`}>
+          <button className="page-link" onClick={() => handlePageChange(currentPage + 1)}>&gt;</button>
+        </li>
+      </ul>
+    </nav>
+  );
+}
+
+export default Pagination;
