+++ src/components/PokemonDetails.jsx
@@ -8,27 +8,8 @@ import PokemonAbilities from './PokemonAbilities';
 import PokemonStats from './PokemonStats';
 import PokemonTypes from './PokemonTypes';
 import { isFavorite } from '../services/favorites.service';
+import typeColors from '../assets/typeColors';
 
-const typeColors = {
-  bug: "#26de81",
-  dragon: "#ffeaa7",
-  electric: "#fed330",
-  fairy: "#FF0069",
-  fighting: "#30336b",
-  fire: "#f0932b",
-  flying: "#81ecec",
-  grass: "#00b894",
-  ground: "#EFB549",
-  ghost: "#a55eea",
-  ice: "#98D8D8",
-  normal: "#95afc0",
-  poison: "#6c5ce7",
-  psychic: "#a29bfe",
-  rock: "#b8a038",
-  steel: "#5a8ea2", 
-  dark: "#705746", 
-  water: "#0190FF",
-};
 
 function PokemonDetails({ show, handleClose, pokemon, onCatch }) {
   const [isCaught, setIsCaught] = useState(false);
