+++ src/components/PokemonList.css
@@ -1,21 +1,62 @@
 .col-custom {
-    flex: 0 0 auto;
-    width: 12.5%; 
-    max-width: 12.5%; 
-    padding-left: 4px; 
-    padding-right: 8px; 
-  }
+  flex: 0 0 auto;
+  width: 12.5%;
+  max-width: 12.5%;
+  padding-left: 4px;
+  padding-right: 8px;
+}
+
 .list-card {
-  width: 165px;
-}  
+  width: 90%;
+}
+
 .card {
   transition: transform 0.2s ease-in-out, box-shadow 0.2s ease-in-out;
 }
+
 .card:hover {
   transform: translateY(-10px) !important;
   box-shadow: 0 8px 16px rgba(0, 0, 0, 0.4) !important;
 }
+
 .clickable-card {
   cursor: pointer; /* Changes the cursor to pointer when hovering */
 }
-  
+
+/* Media queries */
+
+
+@media (max-width: 1200px) {
+  .col-custom {
+    width: 16.66%;
+    max-width: 16.66%;
+  }
+}
+
+@media (max-width: 992px) {
+  .col-custom {
+    width: 20%;
+    max-width: 20%;
+  }
+}
+
+@media (max-width: 768px) {
+  .col-custom {
+    width: 25%;
+    max-width: 25%;
+  }
+}
+
+@media (max-width: 576px) {
+  .col-custom {
+    width: 33.33%;
+    max-width: 33.33%;
+  }
+}
+
+@media (max-width: 400px) {
+  .col-custom {
+    width: 50%;
+    max-width: 50%;
+  }
+}
